chore: update examples
