+++ examples/main/registry/mock.ts
