+++ examples/main/registry/mod.ts
@@ -1,0 +1,3 @@
+import registryHandler from './registry-handler.ts';
+
+export { registryHandler as handleRegistryRequest };
