+++ examples/main/registry/template.json
