+++ examples/private-npm-package-import/.npmrc
@@ -1,0 +1,2 @@
+@meowmeow:registry=http://localhost:9998/_internal/registry/
+//localhost:9998/_internal/registry/:_authToken=MeowMeowToken
