+++ examples/private-npm-package-import/index.js
@@ -1,11 +1,13 @@
 import foobar from "npm:@meowmeow/foobar";
+import isOdd from "npm:is-odd";
 
 console.log(foobar());
 
 export default {
     fetch() {
         return Response.json({
-            meow: typeof foobar
+            meow: typeof foobar,
+            odd: isOdd(1),
         });
     }
 }
