--- examples/private-npm-registry/.npmrc
@@ -1,2 +1,0 @@
-@meowmeow:registry=http://localhost:9997/private-npm-registry-mock/
-//localhost:9997/private-npm-registry-mock/:_authToken=MeowMeowToken
